pyproject.toml
[project]
name = "exploring-agent-auth"
version = "0.1.0"
description = "Eight identity and authorization patterns for AI agents calling tools, in runnable Jupyter notebooks."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28",
    "pyjwt[crypto]>=2.10",
    "python-dotenv>=1.0",
    "openai>=1.50",
    "openai-agents>=0.7",
    "rich>=13.7",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
]

[dependency-groups]
dev = [
    "jupyterlab>=4.2",
    "ipykernel>=6.29",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Make `framework/` importable from notebooks via the editable install that
# `uv sync` performs on this project.
packages = ["framework"]
